move Parser definition to new parser.typing module
